pyproject.toml
[project]
name = "expense_tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.115.8",
    "httpx>=0.28.1",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "pytest>=8.3.4",
    "requests>=2.32.3",
    "sqlalchemy[asyncio]>=2.0.38",
    "sqlmodel>=0.0.22",
]

[tool.uv.workspace]
members = ["expense_tracker"]
